gpu: Remove unnecessary cast before formatting.
With `Display` formatting, floats don't print decimal points if their
values are integers.

+++ all-is-cubes-gpu/src/common/octree_alloc.rs
@@ -604,8 +604,8 @@ impl Fmt<StatusText> for Info {
         write!(
             fmt,
             "{up:3}% used of {ap:3}% allocated of {total_volume}",
-            up = (used_volume as f32 / self.total_volume as f32 * 100.0).ceil() as usize,
-            ap = (allocated_volume as f32 / self.total_volume as f32 * 100.0).ceil() as usize,
+            up = (used_volume as f32 / self.total_volume as f32 * 100.0).ceil(),
+            ap = (allocated_volume as f32 / self.total_volume as f32 * 100.0).ceil(),
         )
     }
 }
